1.2.Flex, Versioning & extra.symfony.require{composer update symfony/*}
composer update symfony/*

If we did not have Symfony Flex installed, we would expect that symfony/form would be updated to 4.4. But... yea! It says:
Restricting packages listed in symfony/symfony to 4.3.*
And... when we find symfony/form, it did upgrade it, but only to the latest 4.3 release - not 4.4. You can also see that it updated a few other libraries that start with symfony/* but that aren't part of the main Symfony code. Flex has no effect on these: they upgrade normally, and that's fine.
So upgrading the "patch" version of Symfony to get bug fixes and security releases is just as simple as running composer update symfony/*. But to upgrade to the next minor version, we need to change the extra.symfony.require key. Except... there will be one other trick. Let's see what it is next.