2.1.Managing Flex, extra.symfony.require & Version Constraints
We just ran:

composer update symfony/*
Thanks to the extra.symfony.require key in our composer.json file:

 104 lines  composer.json
{
    "extra": {
        "symfony": {
            "require": "4.3.*"
        }
    }
}
Which is currently set to 4.3.*, it only upgraded things to the latest 4.3 version - not 4.4. Let's change this to 4.4.*
But wait... why are we upgrading to Symfony 4.4? Isn't this a tutorial about upgrading to Symfony 5? Why not just go straight there? The reason is due to Symfony's, honestly, incredible upgrade policy. Symfony never breaks backwards compatibility for a minor release - like from 4.3 to 4.4. Instead, it deprecates code... and you can see what deprecated code you're using with some special tools. By upgrading to 4.4, we'll be able to see the full list of deprecated things we need to fix. Then we can fix them before upgrading to Symfony 5. We'll see this later.