2.2.Managing Flex, extra.symfony.require & Version Constraints{composer update symfony/*; composer.json Version Constraints for symfony/ Packages: composer show symfony/mailer; Symfony Flex: composer.json Version Formatting; Symfony Flex: symfony.extra.require is a "Soft" Requirement}
Anyways, find your terminal and, once again, run:

composer update symfony/*
Yea! This time it is updating the Symfony packages to 4.4. That was easy!

composer.json Version Constraints for symfony/ Packages
Except... come on... it's never quite that easy. In fact, some Symfony packages did not upgrade. Check it out. Run:

composer show symfony/mailer
Scroll up: woh! This is still on version 4.3! Why?
Open up the composer.json file and find symfony/mailer:

 104 lines  composer.json
{
    "require": {
        "symfony/framework-bundle": "^4.0",
        "symfony/mailer": "4.3.*",
    },
}
Interesting: some packages - like symfony/form or symfony/framework-bundle are set to ^4.0 - which more or less means 4.*. But the symfony/mailer version if 4.3.*.

Symfony Flex: composer.json Version Formatting
There are two things I need to say about this. First, usually when you run compose require some/package, when Composer updates your composer.json file, it uses the "caret" (^) format. That's why you see ^3.0 and ^1.1.
But, when you use Symfony Flex and composer require a Symfony package, it changes that to use the * format - like 4.3.*. That's not a huge deal. In fact, it's almost an accidental feature - but it is nice because the best-practice is typically to control the "minor" version of your Symfony packages - that's the middle number - so that you can upgrade them all at the same time.
But... Flex didn't always do this. That's why, in my project, you see a mixture: some libraries like symfony/form have the "caret" format and other libraries - that were installed more recently like symfony/mailer - use the "star" format.

Symfony Flex: symfony.extra.require is a "Soft" Requirement
The second thing I need to tell you is that the extra.symfony.require config - set to 4.4.* now - is... more of a "suggestion". It doesn't force all Symfony packages to this version. More accurately it says:

  When any symfony/ package is updated, its upgrade will be restricted to a version matching 4.4.*

But if you have a package that is specifically locked to 4.3.*, it won't override that and force it to 4.4.*. That is why symfony/mailer didn't upgrade.