3.1.Upgrading to Symfony 4.4
To upgrade from Symfony 4.3 to 4.4 - that's a "minor" version upgrade - we need to change the extra.symfony.require value to 4.4.* - done! - and update each Symfony package version to that same value.

Updating Versions of Individual Symfony Packages
Let's get to work! I'll start with symfony/asset: change it to 4.4.*. Copy that and start repeating it. I will skip a few packages that start with symfony/ because they are not part of the main Symfony repository - like symfony/flex.
These follow their own release schedules... so they usually have a version that's very different than everything else.
All "packs" - those are the, sort of, "fake" packages that just require other packages for convenience - are another example.
These usually allow pretty much any version of the libraries inside of them - so any Symfony packages will update correctly. If you want more control over the versions, remember that you can run:

composer unpack symfony/orm-pack
When you do that, Flex will remove this line and replace it with the individual packages so you can manage their versions. That's not required, but also not a bad idea.
WebpackEncoreBundle is another example of a package that isn't part of the main repository - you can see that its version is totally different.
Don't forget to also check the require-dev section: there are a bunch here. Including symfony/debug-bundle, which has a funny-looking version because I unpacked it from a debug-pack in one of our courses. And both MakerBundle and MonologBundle are not in the main repository.
If you're not sure, you can search Packagist.org for symfony/symfony. That package lists all of the packages that make up this main repository I keep talking about.