3.2.Upgrading to Symfony 4.4{composer update symfony/*}
Perfect! We have 4.4.* everywhere up here and 4.4.* for extra.symfony.require so that everything matches and we get that performance boost in Composer.

Let's do this! Find your terminal and run:

composer update symfony/*
And... yea! It's upgrading the last few libraries that were previously locked to 4.3. Congratulations! You just upgraded all Symfony packages to 4.4.