3.4.Upgrading to Symfony 4.4{Fixing some require-dev Packages: composer require symfony/dotenv symfony/monolog-bundle easycorp/easy-log-handler}
Re-add the libraries by using that same command, but replacing remove with require and getting rid of the --dev flag:

composer require symfony/dotenv symfony/monolog-bundle easycorp/easy-log-handler
This should add those back under the require section - yep, here is one - and it will reinstall the latest version of their recipes... which means that the recipe could be slightly newer than the one we had before.
This is... accidentally... the first example of upgrading a recipe. Run:

git status
Cool. Should we commit all of these changes? Not so fast. When a recipe is updated, you need to selectively add each change. Let's learn how next.