9.2.Updating the TwigBundle Recipe{Manually Deleting config/routes/dev/twig.yaml}
Manually Deleting config/routes/dev/twig.yaml
We're done! But... wait a second. We expected that the updated recipe would delete the extra config/routes/dev/twig.yaml file... but it didn't. Hmm... is it still in the recipe for some reason? Run:

composer recipes symfony/twig-bundle
Copy the URL to the recipe... and paste it in your browser. Huh. No - there is no config/routes directory at all in here. The file is gone! Why wasn't it deleted?
This is a shortcoming of the recipe update system: it's not smart enough. In a perfect world, it would realize that there used to be a config/routes/dev/twig.yaml file in the old version of the recipe... and since it is not there in the new version, it would delete it. But, that does not happen, at least not yet.
So, we need to delete it manually. This doesn't happen very often, but it is something you should be aware of.

Back at the terminal, run:

git status
one more time - things look good - and let's commit:

git commit -m "updating symfony/twig-bundle"
Nice! Now run:

composer recipes
We're getting close! Let's do any easy one next: let's upgrade symfony/mailer and symfony/sendgrid-mailer.