11.1.phpunit-bridge & routing Recipes
Let's see what recipes we have left:

composer recipes
Updating symfony/phpunit-bridge
Next up is phpunit-bridge. Copy its name and run:

composer recipes:install symfony/phpunit-bridge --force -v
It says it created - probably updated - 5 files. You know the process:

git add -p
Updating .env.test
The first is .env.test with two changes. This fix is a silly typo on the APP_SECRET variable - that's not important - and the second is a new PANTHER_APP_ENV variable. If you're using Symfony Panther - a cool testing tool - then this variable is used to tell Panther which environment to launch. If you're not, then you don't technically need this... but it doesn't hurt anything. You also might install Panther in the future.
As a challenge, let's see if we can just add the second change. To do that, type "s", which means "split". The tool will then try to ask you about each change independently. Say "n" to the first and "y" to the second.

.phpunit.result.cache inside .gitignore
The next update is inside .gitignore: it ignores a new .phpunit.result.cache file. This is why I love updating recipes. Since version 7.3, PHPUnit outputs this file under certain conditions as a way to remember which tests failed. It should be ignored, and this takes care of that. Enter "y".

Updating bin/phpunit
Woh! The next change looks bigger - this is bin/phpunit: a script that this packages adds to help execute PHPUnit. It has a number of subtle updates... and since you've almost definitely not made any custom tweaks to this file, let's add the change.

Updating phpunit.xml.dist
The last update is for phpunit.xml.dist - PHPUnit's configuration file. You may have some customizations here that you want to keep - so be careful. The recipe updates are tiny: these changed from <env to <server - a mostly meaningless change to how environment variables are added - and it looks like something about these two SYMFONY_PHPUNIT variables got tweaked a bit.
Hit "y" to accept this patch. The last change is for symfony.lock - hit "y" for this one too.