12.3.Updating security, translation & validator Recipes{Updating the symfony/validator recipe}
Updating the symfony/validator recipe
We're on a roll!

composer recipes
Oh, so close. Next is the validator recipe. Update it:

composer recipes:install symfony/validator --force -v
And walk through the changes:

git add -p
The first change is in config/packages/validator.yaml: it adds some new config that's commented out. This activates a new validation feature called auto-mapping. It's really cool - and we're going to talk about it later. Hit "y" to add these comments and... yep! This is the symfony.lock file. Press "y" again.

That was easy! Let's commit. Actually, I should have run git status, because this recipe also added a new file. We'll see it in a minute:

git commit -m "updating validator recipe"
Updating webpack-encore-bundle Recipe
Are we done? Run:

composer recipes
We are! Well, there is one more that starts with symfony/: webpack-encore-bundle. But that bundle isn't part of the main Symfony repository... so you can update it now or later. If you're interested, let's update it next. If you're not, skip ahead one chapter to start finding and fixing deprecations.