15.1.Upgrading KnpPaginatorBundle & PHP Platform Version
If you look at the list of deprecations, a bunch of these mention the same problem: some TreeBuilder::root() thing. This is a low-level function that third-party bundles use. And if you dig through the list, this stof_doctrine_extensions comes from StofDoctrineExtensionsBundle... as does most of the other ones - like orm, and mongodb. The last one comes from KnpPaginatorBundle.
So basically... we need to upgrade both KnpPaginatorBundle and StofDoctrineExtensionsBundle.

Upgrading KnpPaginatorBundle the Lazy Way
Let's start with KnpPaginatorBundle... and try to be as lazy as possible. Copy the package name, move over, and run:

composer update knplabs/knp-paginator-bundle
My hope is that a minor upgrade - something like 2.8 to 2.9, which my composer.json version constraint allows - will be enough to fix the deprecation:

 104 lines  composer.json
{
    "require": {
        "knplabs/knp-paginator-bundle": "^2.7",
    },
}
And... absolutely nothing happens. It didn't upgrade the library at all! Boo.

Digging into a Library's Releases
So much for the lazy way out: now we need to do some digging. Google for the bundle and find their GitHub page. PhpStorm tells me that I currently have version 2.8.0. Back on the GitHub page, click on "Releases".
Woh! The latest version is 5.0! And it says:

Added support for Symfony 5

That's what we want! So, to get a version of this library that works with Symfony 5, we need to upgrade to 5.0 of the bundle. Back in composer.json, change the version to ^5.0